+++ Core/Entities/Characters/Enemy/Behavior/Chase.cs
@@ -1,8 +1,6 @@
-﻿using System.Diagnostics;
-using ECS2022_23.Core.World;
-using Microsoft.Xna.Framework;
+﻿using Microsoft.Xna.Framework;
 
-namespace ECS2022_23.Core.Entities.Characters.enemy.enemyBehavior;
+namespace ECS2022_23.Core.Entities.Characters.Enemy.Behavior;
 
 public class Chase : Behavior
 {
