+++ Core/Entities/Characters/Enemy/Behavior/Pathfinding.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using ECS2022_23.Core.World;
 
-namespace ECS2022_23.Core.Entities.Characters.enemy.enemyBehavior;
+namespace ECS2022_23.Core.Entities.Characters.Enemy.Behavior;
 
 public class Pathfinding
 {
