+++ Core/Entities/Characters/Enemy/Behavior/StationaryShooter.cs
@@ -1,6 +1,6 @@
 using Microsoft.Xna.Framework;
 
-namespace ECS2022_23.Core.Entities.Characters.enemy.enemyBehavior;
+namespace ECS2022_23.Core.Entities.Characters.Enemy.Behavior;
 
 public class StationaryShooter : Behavior
 {
