+++ Core/Entities/ProjectileShot.cs
@@ -1,6 +1,6 @@
 ﻿using System.Linq;
 using ECS2022_23.Core.Entities.Characters;
-using ECS2022_23.Core.Entities.Characters.enemy;
+using ECS2022_23.Core.Entities.Characters.Enemy;
 using ECS2022_23.Core.Entities.Items;
 using ECS2022_23.Core.World;
 using ECS2022_23.Enums;
