+++ Core/Escape/Escape.cs
@@ -1,6 +1,5 @@
 using Comora;
 using ECS2022_23.Core.Entities.Characters;
-using ECS2022_23.Core.Entities.Characters.enemy;
 using ECS2022_23.Core.Screens;
 using ECS2022_23.Core.Manager;
 using ECS2022_23.Core.Sound;
