+++ Core/Manager/CombatManager.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using ECS2022_23.Core.Entities;
 using ECS2022_23.Core.Entities.Characters;
-using ECS2022_23.Core.Entities.Characters.enemy;
+using ECS2022_23.Core.Entities.Characters.Enemy;
 using ECS2022_23.Core.Loader;
 using ECS2022_23.Core.Sound;
 using ECS2022_23.Enums;
